+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Snapshot 18w18b
+
+### Added
+
+### Modified
+
+* Typos in sprites for barrels
+
 ## Snapshot 18w18a
 
 * Initial Snapshot Release